move types into new types file
