+++ src/aics-image-viewer/shared/utils/viewerChannelSettings.ts
@@ -1,27 +1,4 @@
-import { ControlPoint } from "@aics/volume-viewer";
 import { OTHER_CHANNEL_KEY, SINGLE_GROUP_CHANNEL_KEY } from "../constants";
-import { ColorArray } from "./colorRepresentations";
-
-/** Settings for a single channel, as stored internally by the app */
-export interface ChannelState {
-  name: string;
-  volumeEnabled: boolean;
-  isosurfaceEnabled: boolean;
-  isovalue: number;
-  colorizeEnabled: boolean;
-  colorizeAlpha: number;
-  opacity: number;
-  color: ColorArray;
-  controlPoints: ControlPoint[];
-}
-
-export type ChannelStateKey = keyof ChannelState;
-export type ChannelSettingUpdater = <K extends ChannelStateKey>(index: number, key: K, value: ChannelState[K]) => void;
-export type MultipleChannelSettingsUpdater = <K extends ChannelStateKey>(
-  indices: number[],
-  key: K,
-  value: ChannelState[K]
-) => void;
 
 /** Settings for a single channel, as passed in via props by App users */
 export interface ViewerChannelSetting {
